(tty_color_get_name_by_index): cleanup format compiler warning.
Found by GCC 6.2.0.

color-internal.c:186:33: error: format '%X' expects argument of type 'unsigned int', but argument 3 has type 'int' [-Werror=format=]

+++ lib/tty/color-internal.c
@@ -183,7 +183,7 @@ tty_color_get_name_by_index (int idx)
         if (idx < 256)
             sprintf (name, "color%d", idx);
         else
-            sprintf (name, "#%06X", idx & 0xFFFFFF);
+            sprintf (name, "#%06X", (unsigned int) idx & 0xFFFFFF);
         return g_intern_string (name);
     }
     return "default";
